Fixing max root so root 0 being unflushed can be differentiated from root0 being flushed

+++ accounts-db/src/accounts_db.rs
@@ -4497,7 +4497,7 @@ impl AccountsDb {
             }
         }
 
-        max_flush_root.inspect(|&root| self.accounts_cache.set_max_flush_root(root));
+        max_flush_root.inspect(|&root| self.accounts_cache.set_max_flush_root(root as i64));
 
         (num_new_roots, num_roots_flushed, flush_stats)
     }
